# dict/api/user/index.ts
/* eslint-disable */
// tslint:disable
/**
 * MICRO SERVICES ENDPOINTS
 * 
 *
 * OpenAPI spec version: 2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import passport from 'passport'
import * as t from './types'
import * as v from '../../validation'
import { Api } from '../../models'

export default function(app: Express, impl: t.UserApi) {
	app.post(
		'/createUser',
		function (req, res) {
			try {
				function __body() {
					const __contentType = req.get('Content-Type')
					const __mimeType = __contentType ? __contentType.replace(/;.*/, '') : undefined

					if (__mimeType === 'application/json') {
						return v.modelApiBODYDATAFromJson('body', req.body)
					}
					console.error(`Invalid request content type: ${__contentType}`)
					throw new Error(`Invalid request content type: ${__contentType}`)
				}

				impl.userCreate(__body()).then(function (response) {
					if (response.status === 201) {
						let body: any
						try {
							body = v.modelApiApiResponseToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in user.userCreate', error)
							res.status(500)
							res.send()
							return
						}

						res.status(201)
						res.send(body)
						return
					}
					if (response.status === 400) {
						let body: any
						try {
							body = v.modelApiApiResponseToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in user.userCreate', error)
							res.status(500)
							res.send()
							return
						}

						res.status(400)
						res.send(body)
						return
					}

					console.log('Unsupported response in user.userCreate', response)
					res.status(500)
					res.send()
				}).catch(function (error) {
					console.error('Unexpected error in user.userCreate', error.stack || error)
					res.status(500)
					res.send()
				})
			} catch (error) {
				/* Catch validation errors */
				res.status(400)
				res.send(error)
			}
		}
	)

	app.delete(
		'/deleteUser/:Topicid',
		function (req, res) {
			try {
				impl.deleteUser(v.parseString('params.Topicid', req.params['Topicid'])).then(function (response) {
					if (response.status === 201) {
						let body: any
						try {
							body = v.modelApiApiResponseToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in user.deleteUser', error)
							res.status(500)
							res.send()
							return
						}

						res.status(201)
						res.send(body)
						return
					}
					if (response.status === 400) {
						let body: any
						try {
							body = v.modelApiApiResponseToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in user.deleteUser', error)
							res.status(500)
							res.send()
							return
						}

						res.status(400)
						res.send(body)
						return
					}

					console.log('Unsupported response in user.deleteUser', response)
					res.status(500)
					res.send()
				}).catch(function (error) {
					console.error('Unexpected error in user.deleteUser', error.stack || error)
					res.status(500)
					res.send()
				})
			} catch (error) {
				/* Catch validation errors */
				res.status(400)
				res.send(error)
			}
		}
	)

	app.get(
		'/getUser',
		function (req, res) {
			try {
				impl.getUser().then(function (response) {
					if (response.status === 201) {
						let body: any
						try {
							body = v.arrayToJson(v.modelApiBODYDATAToJson)('response', response.body)
						} catch (error) {
							console.error('Invalid response body in user.getUser', error)
							res.status(500)
							res.send()
							return
						}

						res.status(201)
						res.send(body)
						return
					}
					if (response.status === 400) {
						let body: any
						try {
							body = v.modelApiApiResponseToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in user.getUser', error)
							res.status(500)
							res.send()
							return
						}

						res.status(400)
						res.send(body)
						return
					}

					console.log('Unsupported response in user.getUser', response)
					res.status(500)
					res.send()
				}).catch(function (error) {
					console.error('Unexpected error in user.getUser', error.stack || error)
					res.status(500)
					res.send()
				})
			} catch (error) {
				/* Catch validation errors */
				res.status(400)
				res.send(error)
			}
		}
	)

	app.put(
		'/UpdateUser/:Topicid',
		function (req, res) {
			try {
				function __body() {
					const __contentType = req.get('Content-Type')
					const __mimeType = __contentType ? __contentType.replace(/;.*/, '') : undefined

					if (__mimeType === 'application/json') {
						return v.modelApiBODYDATAFromJson('body', req.body)
					}
					console.error(`Invalid request content type: ${__contentType}`)
					throw new Error(`Invalid request content type: ${__contentType}`)
				}

				impl.updateUser(v.parseString('params.Topicid', req.params['Topicid']), __body()).then(function (response) {
					if (response.status === 201) {
						let body: any
						try {
							body = v.modelApiApiResponseToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in user.updateUser', error)
							res.status(500)
							res.send()
							return
						}

						res.status(201)
						res.send(body)
						return
					}
					if (response.status === 400) {
						let body: any
						try {
							body = v.modelApiApiResponseToJson('response', response.body)
						} catch (error) {
							console.error('Invalid response body in user.updateUser', error)
							res.status(500)
							res.send()
							return
						}

						res.status(400)
						res.send(body)
						return
					}

					console.log('Unsupported response in user.updateUser', response)
					res.status(500)
					res.send()
				}).catch(function (error) {
					console.error('Unexpected error in user.updateUser', error.stack || error)
					res.status(500)
					res.send()
				})
			} catch (error) {
				/* Catch validation errors */
				res.status(400)
				res.send(error)
			}
		}
	)

}